Refactor: standardize `Select` component paths and extract utility functions ♻️
- Renamed `Select` component files to lowercase (`Select` → `select`) for consistency.
- Moved utility functions (`isEqual`, `include`, `isOption`) to a dedicated `helpers.ts` file.
- Updated all related imports and references to reflect the new paths.
- Adjusted `CubeBelongsTo` and TypeScript types for compatibility with the updated structure.
