+++ resources/js/inertia/components/form/fields/select/types.ts
@@ -5,11 +5,10 @@ export interface Option {
   value: any;
 }
 
-export interface SelectInputProps
-  extends Omit<
-    React.ComponentProps<"input">,
-    "name" | "className" | "value" | "onInput" | "ref" | "onChange"
-  > {}
+export interface SelectInputProps extends Omit<
+  React.ComponentProps<"input">,
+  "name" | "className" | "value" | "onInput" | "ref" | "onChange"
+> {}
 
 export interface IApiSelectProps<TResponse, TData> {
   api: (
@@ -94,13 +93,3 @@ export interface ISelectProps<TData> {
   inputProps?: SelectInputProps;
   required?: boolean;
 }
-
-export const isEqual = (option1: Option, option2: Option): boolean =>
-  (option1.label ?? undefined) == (option2.label ?? undefined) &&
-  (option1.value ?? undefined) == (option2.value ?? undefined);
-
-export const include = (option: Option, selected: Option[]): boolean =>
-  selected.filter((op) => isEqual(op, option)).length > 0;
-
-export const isOption = (object: any): object is Option =>
-  "label" in object && "value" in object;
